--- kurtis/commands/dataset/augment.py
@@ -1,12 +1,0 @@
-# -----------------------------------------------------------------------
-# DISCLAIMER: This code is manually authored and reviewed by humans.
-#             Includes natural human imperfections and creative patterns.
-#             Any imperfections are responsibly owned by its human
-#             creators.
-# -----------------------------------------------------------------------
-import click
-
-
-@click.command(name="augment")
-def command():
-    raise NotImplementedError()
